# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Ho Chi Minh City Food Safety Authority portal.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PublicSummary {
  eligibleFacilities: number;
  selfDeclaredProducts: number;
  registeredProducts: number;
  licensedAdvertising: number;
  testingFacilities: number;
}

export type PublicRecordCategory = typeof PublicRecordCategory[keyof typeof PublicRecordCategory];


export const PublicRecordCategory = {
  'eligible-facilities': 'eligible-facilities',
  'self-declared-products': 'self-declared-products',
  'registered-products': 'registered-products',
  'licensed-advertising': 'licensed-advertising',
  'testing-facilities': 'testing-facilities',
} as const;

export type PublicRecordMetadata = {[key: string]: string};

export interface PublicRecord {
  id: string;
  category: PublicRecordCategory;
  title: string;
  subtitle: string;
  location: string;
  status: string;
  publishedAt: string;
  metadata: PublicRecordMetadata;
}

export interface PublicRecordPage {
  items: PublicRecord[];
  total: number;
  page: number;
  pageSize: number;
}

export interface Attachment {
  name: string;
  kind: string;
  size: number;
  fieldKey?: string;
}

export type ApplicationType = typeof ApplicationType[keyof typeof ApplicationType];


export const ApplicationType = {
  'food-supplier': 'food-supplier',
  'meal-provider': 'meal-provider',
  school: 'school',
} as const;

export type CriteriaAnswerType = typeof CriteriaAnswerType[keyof typeof CriteriaAnswerType];


export const CriteriaAnswerType = {
  text: 'text',
  number: 'number',
  date: 'date',
  'yes-no': 'yes-no',
  select: 'select',
  file: 'file',
} as const;

export interface CriteriaDefinition {
  id: string;
  key: string;
  label: string;
  description: string;
  group: string;
  answerType: CriteriaAnswerType;
  options: string[];
  /** @minimum 0 */
  maxScore: number;
  required: boolean;
  prerequisite: boolean;
  active: boolean;
  order: number;
}

export interface CriteriaSet {
  type: ApplicationType;
  version: string;
  effectiveFrom: string;
  totalScore: number;
  criteria: CriteriaDefinition[];
}

export type CriteriaConfigInputApplyMode = typeof CriteriaConfigInputApplyMode[keyof typeof CriteriaConfigInputApplyMode];


export const CriteriaConfigInputApplyMode = {
  now: 'now',
  scheduled: 'scheduled',
} as const;

export interface CriteriaConfigInput {
  criteria: CriteriaDefinition[];
  effectiveFrom: string;
  applyMode: CriteriaConfigInputApplyMode;
}

export type CriteriaHistoryEntryOldValue = { [key: string]: unknown };

export type CriteriaHistoryEntryNewValue = { [key: string]: unknown };

export interface CriteriaHistoryEntry {
  id: string;
  type: ApplicationType;
  version: string;
  changedAt: string;
  changedBy: string;
  changeType: string;
  summary: string;
  oldValue: CriteriaHistoryEntryOldValue;
  newValue: CriteriaHistoryEntryNewValue;
}

export type ApplicationTypeProperty = typeof ApplicationTypeProperty[keyof typeof ApplicationTypeProperty];


export const ApplicationTypeProperty = {
  'food-supplier': 'food-supplier',
  'meal-provider': 'meal-provider',
  school: 'school',
} as const;

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  pending: 'pending',
  'needs-more-info': 'needs-more-info',
  approved: 'approved',
  warning: 'warning',
  stopped: 'stopped',
  rejected: 'rejected',
} as const;

export type ApplicationData = { [key: string]: unknown };

export type ApplicationScoreBreakdown = {[key: string]: number};

export type ApplicationPrerequisiteResults = {[key: string]: boolean};

export interface Application {
  id: string;
  reference: string;
  type: ApplicationTypeProperty;
  applicantName: string;
  address: string;
  contact: string;
  submittedAt: string;
  status: ApplicationStatus;
  score: number;
  /** @nullable */
  reviewNote: string | null;
  isThirdParty: boolean;
  data: ApplicationData;
  attachments: Attachment[];
  criteriaVersion: string;
  criteriaSnapshot: CriteriaDefinition[];
  scoreBreakdown: ApplicationScoreBreakdown;
  prerequisiteResults: ApplicationPrerequisiteResults;
}

export type ApplicationInputType = typeof ApplicationInputType[keyof typeof ApplicationInputType];


export const ApplicationInputType = {
  'food-supplier': 'food-supplier',
  'meal-provider': 'meal-provider',
  school: 'school',
} as const;

export type ApplicationInputData = { [key: string]: unknown };

export interface ApplicationInput {
  type: ApplicationInputType;
  /** @minLength 1 */
  applicantName: string;
  /** @minLength 1 */
  address: string;
  /** @minLength 1 */
  contact: string;
  data: ApplicationInputData;
  attachments: Attachment[];
  criteriaVersion: string;
  isThirdParty?: boolean;
}

export type ReviewInputAction = typeof ReviewInputAction[keyof typeof ReviewInputAction];


export const ReviewInputAction = {
  approve: 'approve',
  reject: 'reject',
  'needs-more-info': 'needs-more-info',
} as const;

export type ReviewInputCriteriaScores = {[key: string]: number};

export type ReviewInputPrerequisiteResults = {[key: string]: boolean};

export interface ReviewInput {
  action: ReviewInputAction;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  note: string;
  criteriaScores: ReviewInputCriteriaScores;
  prerequisiteResults: ReviewInputPrerequisiteResults;
}

export type AdminSummaryByType = {[key: string]: number};

export interface AdminSummary {
  pending: number;
  needsMoreInfo: number;
  approved: number;
  warning: number;
  stopped: number;
  rejected: number;
  total: number;
  byType: AdminSummaryByType;
}

export interface SupplierOption {
  id: string;
  name: string;
  taxCode: string;
  phone: string;
  verified: boolean;
}

export type ListPublicRecordsParams = {
category?: ListPublicRecordsCategory;
search?: string;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: number;
};

export type ListPublicRecordsCategory = typeof ListPublicRecordsCategory[keyof typeof ListPublicRecordsCategory];


export const ListPublicRecordsCategory = {
  'eligible-facilities': 'eligible-facilities',
  'self-declared-products': 'self-declared-products',
  'registered-products': 'registered-products',
  'licensed-advertising': 'licensed-advertising',
  'testing-facilities': 'testing-facilities',
} as const;

export type ListApplicationsParams = {
status?: ListApplicationsStatus;
type?: ListApplicationsType;
search?: string;
};

export type ListApplicationsStatus = typeof ListApplicationsStatus[keyof typeof ListApplicationsStatus];


export const ListApplicationsStatus = {
  pending: 'pending',
  'needs-more-info': 'needs-more-info',
  approved: 'approved',
  warning: 'warning',
  stopped: 'stopped',
  rejected: 'rejected',
} as const;

export type ListApplicationsType = typeof ListApplicationsType[keyof typeof ListApplicationsType];


export const ListApplicationsType = {
  'food-supplier': 'food-supplier',
  'meal-provider': 'meal-provider',
  school: 'school',
} as const;

export type GetCriteriaParams = {
type: ApplicationType;
version?: string;
};