# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for the Ho Chi Minh City Food Safety Authority portal.
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: public
    description: Public food safety lookup operations
  - name: applications
    description: Registration application operations
  - name: admin
    description: Internal review operations
  - name: criteria
    description: Dynamic registration criteria and form configuration
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /public/summary:
    get:
      operationId: getPublicSummary
      tags: [public]
      summary: Get public portal summary counts
      responses:
        "200":
          description: Public record totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicSummary"
  /public/records:
    get:
      operationId: listPublicRecords
      tags: [public]
      summary: Search publicly available food safety records
      parameters:
        - in: query
          name: category
          schema:
            type: string
            enum: [eligible-facilities, self-declared-products, registered-products, licensed-advertising, testing-facilities]
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: Matching public records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicRecordPage"
  /public/records/{id}:
    get:
      operationId: getPublicRecord
      tags: [public]
      summary: Get one public record
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Public record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicRecord"
        "404":
          description: Record not found
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List submitted applications for the admin workspace
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, needs-more-info, approved, warning, stopped, rejected]
        - in: query
          name: type
          schema:
            type: string
            enum: [food-supplier, meal-provider, school]
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Submitted applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Submit a new registration application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Invalid application data
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get a submitted application
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Application details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Application not found
  /applications/{id}/review:
    patch:
      operationId: reviewApplication
      tags: [admin]
      summary: Approve, reject, or request more information for an application
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Application not found
  /criteria:
    get:
      operationId: getCriteria
      tags: [criteria]
      summary: Get the active criteria set for an application type
      parameters:
        - in: query
          name: type
          required: true
          schema:
            $ref: "#/components/schemas/ApplicationType"
        - in: query
          name: version
          schema:
            type: string
      responses:
        "200":
          description: Criteria set used to generate the registration form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CriteriaSet"
  /criteria/{type}:
    patch:
      operationId: updateCriteria
      tags: [criteria]
      summary: Create a new version of a criteria set
      parameters:
        - in: path
          name: type
          required: true
          schema:
            $ref: "#/components/schemas/ApplicationType"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriteriaConfigInput"
      responses:
        "200":
          description: Updated criteria set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CriteriaSet"
        "400":
          description: The ordinary criteria do not total 100 points
  /criteria/{type}/history:
    get:
      operationId: listCriteriaHistory
      tags: [criteria]
      summary: List criteria configuration change history
      parameters:
        - in: path
          name: type
          required: true
          schema:
            $ref: "#/components/schemas/ApplicationType"
      responses:
        "200":
          description: Criteria configuration history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CriteriaHistoryEntry"
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get admin review dashboard summary
      responses:
        "200":
          description: Review pipeline totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
  /suppliers:
    get:
      operationId: listApprovedSuppliers
      tags: [applications]
      summary: List approved food and meal suppliers for school forms
      responses:
        "200":
          description: Approved suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplierOption"
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    PublicSummary:
      type: object
      properties:
        eligibleFacilities:
          type: integer
        selfDeclaredProducts:
          type: integer
        registeredProducts:
          type: integer
        licensedAdvertising:
          type: integer
        testingFacilities:
          type: integer
      required: [eligibleFacilities, selfDeclaredProducts, registeredProducts, licensedAdvertising, testingFacilities]
    PublicRecord:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
          enum: [eligible-facilities, self-declared-products, registered-products, licensed-advertising, testing-facilities]
        title:
          type: string
        subtitle:
          type: string
        location:
          type: string
        status:
          type: string
        publishedAt:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
      required: [id, category, title, subtitle, location, status, publishedAt, metadata]
    PublicRecordPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PublicRecord"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required: [items, total, page, pageSize]
    Attachment:
      type: object
      properties:
        name:
          type: string
        kind:
          type: string
        size:
          type: integer
        fieldKey:
          type: string
      required: [name, kind, size]
    ApplicationType:
      type: string
      enum: [food-supplier, meal-provider, school]
    CriteriaAnswerType:
      type: string
      enum: [text, number, date, yes-no, select, file]
    CriteriaDefinition:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        label:
          type: string
        description:
          type: string
        group:
          type: string
        answerType:
          $ref: "#/components/schemas/CriteriaAnswerType"
        options:
          type: array
          items:
            type: string
        maxScore:
          type: integer
          minimum: 0
        required:
          type: boolean
        prerequisite:
          type: boolean
        active:
          type: boolean
        order:
          type: integer
      required: [id, key, label, description, group, answerType, options, maxScore, required, prerequisite, active, order]
    CriteriaSet:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/ApplicationType"
        version:
          type: string
        effectiveFrom:
          type: string
        totalScore:
          type: integer
        criteria:
          type: array
          items:
            $ref: "#/components/schemas/CriteriaDefinition"
      required: [type, version, effectiveFrom, totalScore, criteria]
    CriteriaConfigInput:
      type: object
      properties:
        criteria:
          type: array
          items:
            $ref: "#/components/schemas/CriteriaDefinition"
        effectiveFrom:
          type: string
        applyMode:
          type: string
          enum: [now, scheduled]
      required: [criteria, effectiveFrom, applyMode]
    CriteriaHistoryEntry:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/ApplicationType"
        version:
          type: string
        changedAt:
          type: string
        changedBy:
          type: string
        changeType:
          type: string
        summary:
          type: string
        oldValue:
          type: object
          additionalProperties: true
        newValue:
          type: object
          additionalProperties: true
      required: [id, type, version, changedAt, changedBy, changeType, summary, oldValue, newValue]
    Application:
      type: object
      properties:
        id:
          type: string
        reference:
          type: string
        type:
          type: string
          enum: [food-supplier, meal-provider, school]
        applicantName:
          type: string
        address:
          type: string
        contact:
          type: string
        submittedAt:
          type: string
        status:
          type: string
          enum: [pending, needs-more-info, approved, warning, stopped, rejected]
        score:
          type: integer
        reviewNote:
          type: ["string", "null"]
        isThirdParty:
          type: boolean
        data:
          type: object
          additionalProperties: true
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
        criteriaVersion:
          type: string
        criteriaSnapshot:
          type: array
          items:
            $ref: "#/components/schemas/CriteriaDefinition"
        scoreBreakdown:
          type: object
          additionalProperties:
            type: number
        prerequisiteResults:
          type: object
          additionalProperties:
            type: boolean
      required: [id, reference, type, applicantName, address, contact, submittedAt, status, score, reviewNote, isThirdParty, data, attachments, criteriaVersion, criteriaSnapshot, scoreBreakdown, prerequisiteResults]
    ApplicationInput:
      type: object
      properties:
        type:
          type: string
          enum: [food-supplier, meal-provider, school]
        applicantName:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        contact:
          type: string
          minLength: 1
        data:
          type: object
          additionalProperties: true
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
        criteriaVersion:
          type: string
        isThirdParty:
          type: boolean
          default: false
      required: [type, applicantName, address, contact, data, attachments, criteriaVersion]
    ReviewInput:
      type: object
      properties:
        action:
          type: string
          enum: [approve, reject, needs-more-info]
        score:
          type: integer
          minimum: 0
          maximum: 100
        note:
          type: string
        criteriaScores:
          type: object
          additionalProperties:
            type: number
        prerequisiteResults:
          type: object
          additionalProperties:
            type: boolean
      required: [action, score, note, criteriaScores, prerequisiteResults]
    AdminSummary:
      type: object
      properties:
        pending:
          type: integer
        needsMoreInfo:
          type: integer
        approved:
          type: integer
        warning:
          type: integer
        stopped:
          type: integer
        rejected:
          type: integer
        total:
          type: integer
        byType:
          type: object
          additionalProperties:
            type: integer
      required: [pending, needsMoreInfo, approved, warning, stopped, rejected, total, byType]
    SupplierOption:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        taxCode:
          type: string
        phone:
          type: string
        verified:
          type: boolean
      required: [id, name, taxCode, phone, verified]