# lib/api-zod/src/generated/types/applicationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Ho Chi Minh City Food Safety Authority portal.
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationInputData } from './applicationInputData';
import type { ApplicationInputType } from './applicationInputType';
import type { Attachment } from './attachment';

export interface ApplicationInput {
  type: ApplicationInputType;
  /** @minLength 1 */
  applicantName: string;
  /** @minLength 1 */
  address: string;
  /** @minLength 1 */
  contact: string;
  data: ApplicationInputData;
  attachments: Attachment[];
  criteriaVersion: string;
  isThirdParty?: boolean;
}