# lib/api-zod/src/generated/types/criteriaAnswerType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Ho Chi Minh City Food Safety Authority portal.
 * OpenAPI spec version: 0.1.0
 */

export type CriteriaAnswerType = typeof CriteriaAnswerType[keyof typeof CriteriaAnswerType];


export const CriteriaAnswerType = {
  text: 'text',
  number: 'number',
  date: 'date',
  'yes-no': 'yes-no',
  select: 'select',
  file: 'file',
} as const;