# lib/api-zod/src/generated/types/criteriaConfigInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Ho Chi Minh City Food Safety Authority portal.
 * OpenAPI spec version: 0.1.0
 */
import type { CriteriaConfigInputApplyMode } from './criteriaConfigInputApplyMode';
import type { CriteriaDefinition } from './criteriaDefinition';

export interface CriteriaConfigInput {
  criteria: CriteriaDefinition[];
  effectiveFrom: string;
  applyMode: CriteriaConfigInputApplyMode;
}