# lib/api-zod/src/generated/types/criteriaDefinition.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Ho Chi Minh City Food Safety Authority portal.
 * OpenAPI spec version: 0.1.0
 */
import type { CriteriaAnswerType } from './criteriaAnswerType';

export interface CriteriaDefinition {
  id: string;
  key: string;
  label: string;
  description: string;
  group: string;
  answerType: CriteriaAnswerType;
  options: string[];
  /** @minimum 0 */
  maxScore: number;
  required: boolean;
  prerequisite: boolean;
  active: boolean;
  order: number;
}