# lib/api-zod/src/generated/types/criteriaHistoryEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Ho Chi Minh City Food Safety Authority portal.
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationType } from './applicationType';
import type { CriteriaHistoryEntryNewValue } from './criteriaHistoryEntryNewValue';
import type { CriteriaHistoryEntryOldValue } from './criteriaHistoryEntryOldValue';

export interface CriteriaHistoryEntry {
  id: string;
  type: ApplicationType;
  version: string;
  changedAt: string;
  changedBy: string;
  changeType: string;
  summary: string;
  oldValue: CriteriaHistoryEntryOldValue;
  newValue: CriteriaHistoryEntryNewValue;
}