# lib/api-zod/src/generated/types/criteriaSet.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Ho Chi Minh City Food Safety Authority portal.
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationType } from './applicationType';
import type { CriteriaDefinition } from './criteriaDefinition';

export interface CriteriaSet {
  type: ApplicationType;
  version: string;
  effectiveFrom: string;
  totalScore: number;
  criteria: CriteriaDefinition[];
}