# lib/api-zod/src/generated/types/reviewInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Ho Chi Minh City Food Safety Authority portal.
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewInputAction } from './reviewInputAction';
import type { ReviewInputCriteriaScores } from './reviewInputCriteriaScores';
import type { ReviewInputPrerequisiteResults } from './reviewInputPrerequisiteResults';

export interface ReviewInput {
  action: ReviewInputAction;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  note: string;
  criteriaScores: ReviewInputCriteriaScores;
  prerequisiteResults: ReviewInputPrerequisiteResults;
}